chore(db): define redb_value!
